+++ cpp/src/io/orc/orc.cpp
@@ -22,12 +22,20 @@
 namespace cudf {
 namespace io {
 namespace orc {
+
+uint32_t ProtobufReader::read_field_size(const uint8_t *end)
+{
+  auto const size = get<uint32_t>();
+  CUDF_EXPECTS(size <= static_cast<uint32_t>(end - m_cur), "Protobuf parsing out of bounds");
+  return size;
+}
+
 void ProtobufReader::skip_struct_field(int t)
 {
   switch (t) {
-    case PB_TYPE_VARINT: get_u32(); break;
+    case PB_TYPE_VARINT: get<uint32_t>(); break;
     case PB_TYPE_FIXED64: skip_bytes(8); break;
-    case PB_TYPE_FIXEDLEN: skip_bytes(get_u32()); break;
+    case PB_TYPE_FIXEDLEN: skip_bytes(get<uint32_t>()); break;
     case PB_TYPE_FIXED32: skip_bytes(4); break;
     default:
       // printf("invalid type (%d)\n", t);
@@ -35,138 +43,88 @@ void ProtobufReader::skip_struct_field(int t)
   }
 }
 
-bool ProtobufReader::read(PostScript &s, size_t maxlen)
+void ProtobufReader::read(PostScript &s, size_t maxlen)
 {
-  auto op = std::make_tuple(FieldUInt64(1, s.footerLength),
-                            FieldEnum<CompressionKind>(2, s.compression),
-                            FieldUInt32(3, s.compressionBlockSize),
-                            FieldPackedUInt32(4, s.version),
-                            FieldUInt64(5, s.metadataLength),
-                            FieldString(8000, s.magic));
-  return function_builder(s, maxlen, op);
+  auto op = std::make_tuple(make_field_reader(1, s.footerLength),
+                            make_field_reader(2, s.compression),
+                            make_field_reader(3, s.compressionBlockSize),
+                            make_packed_field_reader(4, s.version),
+                            make_field_reader(5, s.metadataLength),
+                            make_field_reader(8000, s.magic));
+  function_builder(s, maxlen, op);
 }
 
-bool ProtobufReader::read(FileFooter &s, size_t maxlen)
+void ProtobufReader::read(FileFooter &s, size_t maxlen)
 {
-  auto op = std::make_tuple(FieldUInt64(1, s.headerLength),
-                            FieldUInt64(2, s.contentLength),
-                            FieldRepeatedStruct(3, s.stripes),
-                            FieldRepeatedStruct(4, s.types),
-                            FieldRepeatedStruct(5, s.metadata),
-                            FieldUInt64(6, s.numberOfRows),
-                            FieldRepeatedStructBlob(7, s.statistics),
-                            FieldUInt32(8, s.rowIndexStride));
-  return function_builder(s, maxlen, op);
+  auto op = std::make_tuple(make_field_reader(1, s.headerLength),
+                            make_field_reader(2, s.contentLength),
+                            make_field_reader(3, s.stripes),
+                            make_field_reader(4, s.types),
+                            make_field_reader(5, s.metadata),
+                            make_field_reader(6, s.numberOfRows),
+                            make_raw_field_reader(7, s.statistics),
+                            make_field_reader(8, s.rowIndexStride));
+  function_builder(s, maxlen, op);
 }
 
-bool ProtobufReader::read(StripeInformation &s, size_t maxlen)
+void ProtobufReader::read(StripeInformation &s, size_t maxlen)
 {
-  auto op = std::make_tuple(FieldUInt64(1, s.offset),
-                            FieldUInt64(2, s.indexLength),
-                            FieldUInt64(3, s.dataLength),
-                            FieldUInt32(4, s.footerLength),
-                            FieldUInt32(5, s.numberOfRows));
-  return function_builder(s, maxlen, op);
+  auto op = std::make_tuple(make_field_reader(1, s.offset),
+                            make_field_reader(2, s.indexLength),
+                            make_field_reader(3, s.dataLength),
+                            make_field_reader(4, s.footerLength),
+                            make_field_reader(5, s.numberOfRows));
+  function_builder(s, maxlen, op);
 }
 
-bool ProtobufReader::read(SchemaType &s, size_t maxlen)
+void ProtobufReader::read(SchemaType &s, size_t maxlen)
 {
-  auto op = std::make_tuple(FieldEnum<TypeKind>(1, s.kind),
-                            FieldPackedUInt32(2, s.subtypes),
-                            FieldRepeatedString(3, s.fieldNames),
-                            FieldUInt32(4, s.maximumLength),
-                            FieldUInt32(5, s.precision),
-                            FieldUInt32(6, s.scale));
-  return function_builder(s, maxlen, op);
+  auto op = std::make_tuple(make_field_reader(1, s.kind),
+                            make_packed_field_reader(2, s.subtypes),
+                            make_field_reader(3, s.fieldNames),
+                            make_field_reader(4, s.maximumLength),
+                            make_field_reader(5, s.precision),
+                            make_field_reader(6, s.scale));
+  function_builder(s, maxlen, op);
 }
 
-bool ProtobufReader::read(UserMetadataItem &s, size_t maxlen)
+void ProtobufReader::read(UserMetadataItem &s, size_t maxlen)
 {
-  auto op = std::make_tuple(FieldString(1, s.name), FieldString(2, s.value));
-  return function_builder(s, maxlen, op);
+  auto op = std::make_tuple(make_field_reader(1, s.name), make_field_reader(2, s.value));
+  function_builder(s, maxlen, op);
 }
 
-bool ProtobufReader::read(StripeFooter &s, size_t maxlen)
+void ProtobufReader::read(StripeFooter &s, size_t maxlen)
 {
-  auto op = std::make_tuple(FieldRepeatedStruct(1, s.streams),
-                            FieldRepeatedStruct(2, s.columns),
-                            FieldString(3, s.writerTimezone));
-  return function_builder(s, maxlen, op);
+  auto op = std::make_tuple(make_field_reader(1, s.streams),
+                            make_field_reader(2, s.columns),
+                            make_field_reader(3, s.writerTimezone));
+  function_builder(s, maxlen, op);
 }
 
-bool ProtobufReader::read(Stream &s, size_t maxlen)
+void ProtobufReader::read(Stream &s, size_t maxlen)
 {
   auto op = std::make_tuple(
-    FieldEnum<StreamKind>(1, s.kind), FieldUInt32(2, s.column), FieldUInt64(3, s.length));
-  return function_builder(s, maxlen, op);
+    make_field_reader(1, s.kind), make_field_reader(2, s.column), make_field_reader(3, s.length));
+  function_builder(s, maxlen, op);
 }
 
-bool ProtobufReader::read(ColumnEncoding &s, size_t maxlen)
+void ProtobufReader::read(ColumnEncoding &s, size_t maxlen)
 {
-  auto op =
-    std::make_tuple(FieldEnum<ColumnEncodingKind>(1, s.kind), FieldUInt32(2, s.dictionarySize));
-  return function_builder(s, maxlen, op);
+  auto op = std::make_tuple(make_field_reader(1, s.kind), make_field_reader(2, s.dictionarySize));
+  function_builder(s, maxlen, op);
 }
 
-bool ProtobufReader::read(StripeStatistics &s, size_t maxlen)
+void ProtobufReader::read(StripeStatistics &s, size_t maxlen)
 {
-  auto op = std::make_tuple(FieldRepeatedStructBlob(1, s.colStats));
-  return function_builder(s, maxlen, op);
+  auto op = std::make_tuple(make_raw_field_reader(1, s.colStats));
+  function_builder(s, maxlen, op);
 }
 
-bool ProtobufReader::read(Metadata &s, size_t maxlen)
+void ProtobufReader::read(Metadata &s, size_t maxlen)
 {
-  auto op = std::make_tuple(FieldRepeatedStruct(1, s.stripeStats));
-  return function_builder(s, maxlen, op);
-}
-
-// return the column name
-std::string FileFooter::GetColumnName(uint32_t column_id)
-{
-  std::string s       = "";
-  uint32_t parent_idx = column_id, idx, field_idx;
-  do {
-    idx        = parent_idx;
-    parent_idx = (idx < types.size()) ? (uint32_t)types[idx].parent_idx : ~0;
-    field_idx  = (parent_idx < types.size()) ? (uint32_t)types[idx].field_idx : ~0;
-    if (parent_idx >= types.size()) break;
-    if (field_idx < types[parent_idx].fieldNames.size()) {
-      if (s.length() > 0)
-        s = types[parent_idx].fieldNames[field_idx] + "." + s;
-      else
-        s = types[parent_idx].fieldNames[field_idx];
-    }
-  } while (parent_idx != idx);
-  // If we have no name (root column), generate a name
-  if (s.length() == 0) { s = "col" + std::to_string(column_id); }
-  return s;
-}
-
-// Initializes the parent_idx field in the schema
-bool ProtobufReader::InitSchema(FileFooter &ff)
-{
-  int32_t schema_size = (int32_t)ff.types.size();
-  for (int32_t i = 0; i < schema_size; i++) {
-    int32_t num_children = (int32_t)ff.types[i].subtypes.size();
-    if (ff.types[i].parent_idx == -1)  // Not initialized
-    {
-      ff.types[i].parent_idx = i;  // set root node as its own parent
-    }
-    for (int32_t j = 0; j < num_children; j++) {
-      uint32_t column_id = ff.types[i].subtypes[j];
-      if (column_id <= (uint32_t)i || column_id >= (uint32_t)schema_size) {
-        // Invalid column id (or at least not a schema index)
-        return false;
-      }
-      if (ff.types[column_id].parent_idx != -1) {
-        // Same node referenced twice
-        return false;
-      }
-      ff.types[column_id].parent_idx = i;
-      ff.types[column_id].field_idx  = j;
-    }
-  }
-  return true;
+  auto op = std::make_tuple(make_field_reader(1, s.stripeStats));
+  function_builder(s, maxlen, op);
 }
 
 /**
@@ -320,6 +278,7 @@ OrcDecompressor::OrcDecompressor(CompressionKind kind, uint32_t blockSize)
   if (kind != NONE) {
     int stream_type = IO_UNCOMP_STREAM_TYPE_INFER;  // Will be treated as invalid
     switch (kind) {
+      case NONE: break;
       case ZLIB:
         stream_type    = IO_UNCOMP_STREAM_TYPE_INFLATE;
         m_log2MaxRatio = 11;  // < 2048:1
@@ -400,6 +359,181 @@ const uint8_t *OrcDecompressor::Decompress(const uint8_t *srcBytes, size_t srcLe
   }
   *dstLen = dst_length;
   return m_buf.data();
+}
+
+metadata::metadata(datasource *const src) : source(src)
+{
+  const auto len         = source->size();
+  const auto max_ps_size = std::min(len, static_cast<size_t>(256));
+
+  // Read uncompressed postscript section (max 255 bytes + 1 byte for length)
+  auto buffer            = source->host_read(len - max_ps_size, max_ps_size);
+  const size_t ps_length = buffer->data()[max_ps_size - 1];
+  const uint8_t *ps_data = &buffer->data()[max_ps_size - ps_length - 1];
+  ProtobufReader(ps_data, ps_length).read(ps);
+  CUDF_EXPECTS(ps.footerLength + ps_length < len, "Invalid footer length");
+
+  // If compression is used, the rest of the metadata is compressed
+  // If no compressed is used, the decompressor is simply a pass-through
+  decompressor = std::make_unique<OrcDecompressor>(ps.compression, ps.compressionBlockSize);
+
+  // Read compressed filefooter section
+  buffer           = source->host_read(len - ps_length - 1 - ps.footerLength, ps.footerLength);
+  size_t ff_length = 0;
+  auto ff_data     = decompressor->Decompress(buffer->data(), ps.footerLength, &ff_length);
+  ProtobufReader(ff_data, ff_length).read(ff);
+  CUDF_EXPECTS(get_num_columns() > 0, "No columns found");
+
+  // Read compressed metadata section
+  buffer =
+    source->host_read(len - ps_length - 1 - ps.footerLength - ps.metadataLength, ps.metadataLength);
+  size_t md_length = 0;
+  auto md_data     = decompressor->Decompress(buffer->data(), ps.metadataLength, &md_length);
+  orc::ProtobufReader(md_data, md_length).read(md);
+}
+
+std::vector<metadata::OrcStripeInfo> metadata::select_stripes(const std::vector<size_type> &stripes,
+                                                              size_type &row_start,
+                                                              size_type &row_count)
+{
+  std::vector<OrcStripeInfo> selection;
+
+  if (!stripes.empty()) {
+    size_t stripe_rows = 0;
+    for (const auto &stripe_idx : stripes) {
+      CUDF_EXPECTS(stripe_idx >= 0 && stripe_idx < get_num_stripes(), "Invalid stripe index");
+      selection.emplace_back(&ff.stripes[stripe_idx], nullptr);
+      stripe_rows += ff.stripes[stripe_idx].numberOfRows;
+    }
+    // row_start is 0 if stripes are set. If this is not true anymore, then
+    // row_start needs to be subtracted to get the correct row_count
+    CUDF_EXPECTS(row_start == 0, "Start row index should be 0");
+    row_count = static_cast<size_type>(stripe_rows);
+  } else {
+    row_start = std::max(row_start, 0);
+    if (row_count < 0) {
+      row_count = static_cast<size_type>(
+        std::min<size_t>(get_total_rows() - row_start, std::numeric_limits<size_type>::max()));
+    } else {
+      row_count = static_cast<size_type>(std::min<size_t>(get_total_rows() - row_start, row_count));
+    }
+    CUDF_EXPECTS(row_count >= 0 && row_start >= 0, "Negative row count or starting row");
+    CUDF_EXPECTS(
+      !(row_start > 0 && (row_count > (std::numeric_limits<size_type>::max() - row_start))),
+      "Summation of starting row index and number of rows would cause overflow");
+
+    size_type stripe_skip_rows = 0;
+    for (size_t i = 0, count = 0; i < ff.stripes.size(); ++i) {
+      count += ff.stripes[i].numberOfRows;
+      if (count > static_cast<size_t>(row_start)) {
+        if (selection.empty()) {
+          stripe_skip_rows =
+            static_cast<size_type>(row_start - (count - ff.stripes[i].numberOfRows));
+        }
+        selection.emplace_back(&ff.stripes[i], nullptr);
+      }
+      if (count >= static_cast<size_t>(row_start) + static_cast<size_t>(row_count)) { break; }
+    }
+    row_start = stripe_skip_rows;
+  }
+
+  // Read each stripe's stripefooter metadata
+  if (not selection.empty()) {
+    stripefooters.resize(selection.size());
+    for (size_t i = 0; i < selection.size(); ++i) {
+      const auto stripe         = selection[i].first;
+      const auto sf_comp_offset = stripe->offset + stripe->indexLength + stripe->dataLength;
+      const auto sf_comp_length = stripe->footerLength;
+      CUDF_EXPECTS(sf_comp_offset + sf_comp_length < source->size(), "Invalid stripe information");
+
+      const auto buffer = source->host_read(sf_comp_offset, sf_comp_length);
+      size_t sf_length  = 0;
+      auto sf_data      = decompressor->Decompress(buffer->data(), sf_comp_length, &sf_length);
+      ProtobufReader(sf_data, sf_length).read(stripefooters[i]);
+      selection[i].second = &stripefooters[i];
+    }
+  }
+
+  return selection;
+}
+
+std::vector<int> metadata::select_columns(std::vector<std::string> use_names,
+                                          bool &has_timestamp_column)
+{
+  std::vector<int> selection;
+
+  if (not use_names.empty()) {
+    int index = 0;
+    for (const auto &use_name : use_names) {
+      for (int i = 0; i < get_num_columns(); ++i, ++index) {
+        if (index >= get_num_columns()) { index = 0; }
+        if (get_column_name(index) == use_name) {
+          selection.emplace_back(index);
+          if (ff.types[index].kind == orc::TIMESTAMP) { has_timestamp_column = true; }
+          index++;
+          break;
+        }
+      }
+    }
+  } else {
+    // For now, only select all leaf nodes
+    for (int i = 0; i < get_num_columns(); ++i) {
+      if (ff.types[i].subtypes.empty()) {
+        selection.emplace_back(i);
+        if (ff.types[i].kind == orc::TIMESTAMP) { has_timestamp_column = true; }
+      }
+    }
+  }
+  CUDF_EXPECTS(selection.size() > 0, "Filtered out all columns");
+
+  return selection;
+}
+
+void metadata::init_column_names()
+{
+  auto const schema_idxs = get_schema_indexes();
+  auto const &types      = ff.types;
+  for (int32_t col_id = 0; col_id < get_num_columns(); ++col_id) {
+    std::string col_name;
+    uint32_t parent_idx = col_id;
+    uint32_t idx        = col_id;
+    do {
+      idx        = parent_idx;
+      parent_idx = (idx < types.size()) ? static_cast<uint32_t>(schema_idxs[idx].parent) : ~0;
+      if (parent_idx >= types.size()) break;
+
+      auto const field_idx =
+        (parent_idx < types.size()) ? static_cast<uint32_t>(schema_idxs[idx].field) : ~0;
+      if (field_idx < types[parent_idx].fieldNames.size()) {
+        col_name =
+          types[parent_idx].fieldNames[field_idx] + (col_name.empty() ? "" : ("." + col_name));
+      }
+    } while (parent_idx != idx);
+    // If we have no name (root column), generate a name
+    column_names.push_back(col_name.empty() ? "col" + std::to_string(col_id) : col_name);
+  }
+}
+
+std::vector<metadata::schema_indexes> metadata::get_schema_indexes() const
+{
+  std::vector<schema_indexes> result(ff.types.size());
+
+  auto const schema_size = static_cast<uint32_t>(result.size());
+  for (uint32_t i = 0; i < schema_size; i++) {
+    auto const &subtypes    = ff.types[i].subtypes;
+    auto const num_children = static_cast<uint32_t>(subtypes.size());
+    if (result[i].parent == -1) {  // Not initialized
+      result[i].parent = i;        // set root node as its own parent
+    }
+    for (uint32_t j = 0; j < num_children; j++) {
+      auto const column_id = subtypes[j];
+      CUDF_EXPECTS(column_id > i && column_id < schema_size, "Invalid column id");
+      CUDF_EXPECTS(result[column_id].parent == -1, "Same node referenced twice");
+      result[column_id].parent = i;
+      result[column_id].field  = j;
+    }
+  }
+  return result;
 }
 
 }  // namespace orc